+++ src/updateFunctions/updateAccsRepay.js
@@ -1,4 +1,4 @@
-import { findDeleteOrDecrease } from './updateFunctions'
+import { findDeleteOrDecrease } from "./findDeleteOrDecrease";
 export const updateAccsRepay = (fromAcc, forAcc, transactionTyp, amount, message, state, prev, date) => {
     return state.accounts.map((acc) => {
         if (acc.username === fromAcc) {
